Rename files as sugested
