+++ xsd-parser-types/src/xml/base64.rs
