+++ xsd-parser-types/src/xml/hex.rs
