+++ xsd-parser-types/src/xml/mod.rs
@@ -4,9 +4,9 @@
 
 mod any_simple_type;
 mod attributes;
-mod base64binary;
+mod base64;
 mod element;
-mod hexbinary;
+mod hex;
 mod mixed;
 mod namespace_scope;
 mod namespaces;
@@ -15,8 +15,8 @@ mod qname;
 mod text;
 mod value;
 
-pub use self::base64binary::{Base64Binary as Base64BinaryBytes, Base64String};
-pub use self::hexbinary::{HexBinary, HexString};
+pub use self::base64::{Base64Binary as Base64BinaryBytes, Base64String};
+pub use self::hex::{HexBinary, HexString};
 
 pub use self::any_simple_type::{AnySimpleType, Base64Binary, Decimal, Integer, Unsigned};
 pub use self::attributes::{
